Extract public-ui and rename mascot off the plugin prefix.
Shared chrome (fold, count, bool box, anchor menu, emoji board) lives in
@biu/public-ui. The mascot library is @biu/public-mascot so it is obvious
these packages are not Cordis plugins.
