+++ docs/plugin-packages.md
@@ -4,7 +4,7 @@
 
 `host/index.ts` 只做：建 Context、读 `host` 表、`plugin()`。json 解析在 `@biu/host-plugin-loader`。`hub` 是 `@biu/host-hub`，写在 `host` 表最后一项（它再去挂 `plugins` 表）。
 
-插件包内部源码一律 `src/host/` 与/或 `src/web/`（`type-*` 仍用 `src/`，它们不是插件）。
+插件包内部源码一律 `src/host/` 与/或 `src/web/`（`type-*` / `public-*` 仍用各自目录，它们不是插件）。
 
 ## 三张表
 
@@ -16,11 +16,12 @@
 
 ## `packages/` 前缀
 
-`ls packages` 按字母分成五类：
+`ls packages` 按字母分成六类：
 
 - **`type-*`**：会被很多包直接 `import` 的契约（纯类型 + 纯函数）。例如 `@biu/type-session`、`@biu/type-agent-loop`、`@biu/type-http`、`@biu/type-slots`。不是插件，不进 json。`currentSessionId` / `runWithSession` 是 ALS 运行时，在 `@biu/host-sessions/scope`。
+- **`public-*`**：共享库 / 公共组件，**不属于 Cordis 插件体系**，不要写进 json。例如 `@biu/public-mascot`（吉祥物）、`@biu/public-ui`（侧栏折叠、计数、勾选框、锚点菜单、emoji 面板）。源码不必拆 `./host` / `./web` 入口。
 - **`host-*`**：host 内核插件（无浏览器入口）。源码在 `src/host/`。
-- **`web-*`**：web 内核插件；`web-mascot` 是共享库，不必进 json。源码在 `src/web/`。
+- **`web-*`**：web 内核插件（壳上的 slots / dock / app-modules / shell 等）。源码在 `src/web/`。
 - **`core-*`**：基础能力（File System / Plugin System / Task System），同一目录，**`exports` 必须把 `./host` 与 `./web` 分开**。json 写 `"package": "@biu/core-file-system/host", "web": "@biu/core-file-system/web"`。
 - **`cap-*`**：能力插件，同一目录，**`exports` 必须把 `./host` 与 `./web` 分开**（禁止一个入口同时带 Node + React）。json 写 `"package": "@biu/cap-chat/host", "web": "@biu/cap-chat/web"`。
