+++ packages/cap-chat/src/web/live-dispatch-table.tsx
@@ -2,8 +2,8 @@ import { memo } from 'react'
 import { CheckCircleIcon, MinusCircleIcon, XCircleIcon, ArrowPathIcon } from '@heroicons/react/16/solid'
 import type { TrajectoryUsage } from '@biu/web-session-view'
 import type { DispatchedTaskRow } from '@biu/web-session-view'
-import { SidebarMascot } from '@biu/web-mascot'
-import { resolveSessionMascot } from '@biu/web-mascot'
+import { SidebarMascot } from '@biu/public-mascot'
+import { resolveSessionMascot } from '@biu/public-mascot'
 import { FolderGlyph } from '@biu/web-session-view/folder-glyph'
 import { UsageInline } from './usage-inline.tsx'
