+++ packages/core-file-system/package.json
@@ -10,6 +10,7 @@
   },
   "dependencies": {
     "@biu/type-file-system": "0.1.0",
+    "@biu/public-ui": "0.1.0",
     "@biu/web-session-view": "0.1.0",
     "@biu/web-snapshot": "0.1.0"
   },
