+++ packages/core-file-system/src/web/controls.tsx
@@ -1,6 +1,7 @@
-import { useEffect, useLayoutEffect, useRef, useState, type ReactNode } from 'react'
+import { useEffect, useRef, useState } from 'react'
 import { createPortal } from 'react-dom'
 import { CheckCircleIcon, ChevronDownIcon, MagnifyingGlassIcon, XMarkIcon } from '@heroicons/react/16/solid'
+import { AnchorMenu } from '@biu/public-ui'
 
 function useDismiss(open: boolean, onClose: () => void, ref: { current: HTMLElement | null }) {
   useEffect(() => {
@@ -11,60 +12,6 @@ function useDismiss(open: boolean, onClose: () => void, ref: { current: HTMLElem
     document.addEventListener('mousedown', onDown)
     return () => document.removeEventListener('mousedown', onDown)
   }, [open, onClose, ref])
-}
-
-function SelectMenu({
-  anchor,
-  onClose,
-  children,
-}: {
-  anchor: HTMLElement | null
-  onClose: () => void
-  children: ReactNode
-}) {
-  const menuRef = useRef<HTMLDivElement>(null)
-  const [box, setBox] = useState({ top: 0, left: 0, width: 200 })
-
-  useLayoutEffect(() => {
-    if (!anchor) return
-    const place = () => {
-      const rect = anchor.getBoundingClientRect()
-      const width = Math.max(220, rect.width)
-      const left = Math.min(rect.left, Math.max(8, window.innerWidth - width - 8))
-      const top = rect.bottom + 4
-      setBox({ top, left, width })
-    }
-    place()
-    window.addEventListener('resize', place)
-    window.addEventListener('scroll', place, true)
-    return () => {
-      window.removeEventListener('resize', place)
-      window.removeEventListener('scroll', place, true)
-    }
-  }, [anchor])
-
-  useEffect(() => {
-    const onDown = (event: MouseEvent) => {
-      const target = event.target as Node
-      if (menuRef.current?.contains(target) || anchor?.contains(target)) return
-      onClose()
-    }
-    document.addEventListener('mousedown', onDown)
-    return () => document.removeEventListener('mousedown', onDown)
-  }, [anchor, onClose])
-
-  if (!anchor) return null
-  return createPortal(
-    <div
-      ref={menuRef}
-      className="fsdb-cellselect-menu"
-      role="listbox"
-      style={{ position: 'fixed', top: box.top, left: box.left, width: box.width, zIndex: 80 }}
-    >
-      {children}
-    </div>,
-    document.body,
-  )
 }
 
 export function CellSelect({
@@ -121,7 +68,7 @@ export function CellSelect({
         {variant === 'field' ? <ChevronDownIcon aria-hidden className="size-[14px] fsdb-cellselect-caret" /> : null}
       </button>
       {open ? (
-        <SelectMenu anchor={triggerRef.current} onClose={close}>
+        <AnchorMenu anchor={triggerRef.current} onClose={close}>
           <label className="fsdb-cellselect-search">
             <MagnifyingGlassIcon aria-hidden className="size-[14px]" />
             <input
@@ -157,7 +104,7 @@ export function CellSelect({
             ))}
             {!filtered.length ? <div className="fsdb-cellselect-empty">没有匹配项</div> : null}
           </div>
-        </SelectMenu>
+        </AnchorMenu>
       ) : null}
     </div>
   )
