+++ packages/core-file-system/src/web/data-sidebar.tsx
@@ -1,5 +1,6 @@
-import { memo, useEffect, useLayoutEffect, useMemo, useState, useSyncExternalStore, type ReactNode } from 'react'
+import { memo, useEffect, useLayoutEffect, useMemo, useState, useSyncExternalStore } from 'react'
 import { createPortal } from 'react-dom'
+import { ChatCount, RecordEmojiBoard, SidebarFold } from '@biu/public-ui'
 import {
   ChevronDoubleLeftIcon,
   ChevronDownIcon,
@@ -47,105 +48,9 @@ import { TableGlyph, ViewModeGlyph } from './nav-glyphs.tsx'
 const SIDEBAR_BRAND_GRADIENT =
   'linear-gradient(105deg, color-mix(in srgb, #0066B0 42%, var(--dsw-hover)), color-mix(in srgb, #5B3E90 40%, var(--dsw-hover)) 52%, color-mix(in srgb, #E22726 42%, var(--dsw-hover)))'
 
-const RECORD_EMOJI_PRESETS = ['⭐', '🔥', '✅', '📌', '💡', '🎯', '📦', '🧩', '📄', '⚡']
-
-export function RecordEmojiBoard({
-  anchor,
-  draft,
-  onDraft,
-  onPick,
-  onClear,
-  onClose,
-}: {
-  anchor: HTMLElement
-  draft: string
-  onDraft: (next: string) => void
-  onPick: (emoji: string) => void
-  onClear: () => void
-  onClose: () => void
-}) {
-  const [pos, setPos] = useState({ left: 0, top: 0 })
-  useLayoutEffect(() => {
-    const place = () => {
-      const box = anchor.getBoundingClientRect()
-      const width = 168
-      setPos({
-        left: Math.min(box.left, Math.max(8, window.innerWidth - width - 8)),
-        top: box.bottom + 4,
-      })
-    }
-    place()
-    window.addEventListener('resize', place)
-    window.addEventListener('scroll', place, true)
-    return () => {
-      window.removeEventListener('resize', place)
-      window.removeEventListener('scroll', place, true)
-    }
-  }, [anchor])
-  useEffect(() => {
-    const onPointer = (event: MouseEvent) => {
-      const target = event.target
-      if (!(target instanceof Node)) return
-      if (anchor.contains(target)) return
-      if (target instanceof Element && target.closest('.fsdb-emoji-picker')) return
-      onClose()
-    }
-    document.addEventListener('mousedown', onPointer)
-    return () => document.removeEventListener('mousedown', onPointer)
-  }, [anchor, onClose])
-  if (typeof document === 'undefined') return null
-  return createPortal(
-    <div
-      className="fsdb-emoji-picker is-fixed"
-      data-biu-ignore
-      role="dialog"
-      aria-label="选择图标"
-      style={{ left: pos.left, top: pos.top }}
-      onMouseDown={(event) => event.stopPropagation()}
-      onClick={(event) => event.stopPropagation()}
-    >
-      <div className="fsdb-emoji-picker-presets">
-        {RECORD_EMOJI_PRESETS.map((item) => (
-          <button key={item} type="button" className="fsdb-emoji-picker-item" onClick={() => onPick(item)}>
-            {item}
-          </button>
-        ))}
-      </div>
-      <input
-        className="fsdb-emoji-picker-input"
-        value={draft}
-        placeholder="输入 emoji"
-        maxLength={8}
-        onChange={(event) => onDraft(event.target.value)}
-        onKeyDown={(event) => {
-          if (event.key === 'Enter') {
-            event.preventDefault()
-            onPick(draft)
-          }
-          if (event.key === 'Escape') onClose()
-        }}
-      />
-      <button type="button" className="fsdb-emoji-picker-clear" onClick={onClear}>
-        恢复默认
-      </button>
-    </div>,
-    document.body,
-  )
-}
-
 type PreviewState = { items: DbRecord[]; total: number; loading: boolean; error: string }
 
 const previewCache = new Map<string, { items: DbRecord[]; total: number }>()
-
-function SidebarFold({ open, children, className }: { open: boolean; children: ReactNode; className?: string }) {
-  return (
-    <div className={`sidebar-fold${open ? ' is-open' : ''}${className ? ` ${className}` : ''}`} aria-hidden={!open}>
-      <div className="sidebar-fold-inner" inert={!open || undefined}>
-        {children}
-      </div>
-    </div>
-  )
-}
 
 function ViewRecordPreview({
   path,
@@ -309,15 +214,6 @@ function ViewRecordPreview({
       ) : null}
       {capped ? <div className="fsdb-view-preview-hint">侧栏最多预览 {SIDEBAR_PREVIEW_MAX} 条，完整数据在主区</div> : null}
     </div>
-  )
-}
-
-function ChatCount({ count }: { count: number | undefined }) {
-  if (count == null) return null
-  return (
-    <span className="sidebar-chat-count" title={`${count} 条`}>
-      <span className="sidebar-chat-count-num">{count}</span>
-    </span>
   )
 }
