+++ packages/core-file-system/src/web/inspector-toggle.test.ts
@@ -18,12 +18,14 @@ test('data sidebar brand sits left with a collapse control on the right', () => 
 
 test('user and system collection sections fold independently', () => {
   const sidebar = readFileSync(resolve(import.meta.dirname, './data-sidebar.tsx'), 'utf8')
+  const fold = readFileSync(resolve(import.meta.dirname, '../../../public-ui/src/sidebar-fold.tsx'), 'utf8')
   assert.match(sidebar, /const \[userOpen, setUserOpen\] = useState\(true\)/)
   assert.match(sidebar, /const \[systemOpen, setSystemOpen\] = useState\(true\)/)
   assert.match(sidebar, /onClick=\{\(\) => setUserOpen\(\(prev\) => !prev\)\}/)
   assert.match(sidebar, /onClick=\{\(\) => setSystemOpen\(\(prev\) => !prev\)\}/)
-  assert.match(sidebar, /function SidebarFold/)
-  assert.match(sidebar, /className=\{\`sidebar-fold\$\{open \? ' is-open' : ''\}/)
+  assert.match(sidebar, /<SidebarFold/)
+  assert.match(fold, /function SidebarFold/)
+  assert.match(fold, /className=\{\`sidebar-fold\$\{open \? ' is-open' : ''\}/)
   assert.doesNotMatch(sidebar, /sidebar-session-list min-w-0 \$\{open \? '' : 'hidden'\}/)
   assert.doesNotMatch(sidebar, /\{userOpen \?/)
   assert.doesNotMatch(sidebar, /\{systemOpen \?/)
@@ -71,8 +73,9 @@ test('sidebar records paint detail immediately without reloading the view', () =
   assert.match(browser, /!detailId \?/)
   assert.match(sidebar, /onOpenRecord\?\.\(row\.id, row\)/)
   assert.match(sidebar, /event.stopPropagation\(\)/)
-  assert.match(sidebar, /function RecordEmojiBoard/)
-  assert.match(sidebar, /className="fsdb-emoji-picker is-fixed"/)
+  const emoji = readFileSync(resolve(import.meta.dirname, '../../../public-ui/src/emoji-board.tsx'), 'utf8')
+  assert.match(emoji, /function RecordEmojiBoard/)
+  assert.match(emoji, /className="fsdb-emoji-picker is-fixed"/)
   assert.match(sidebar, /fsdb-record-emoji/)
   assert.match(css, /\.fsdb-emoji-picker \{[^}]*border:\s*0/s)
   assert.match(css, /\.fsdb-emoji-picker \{[^}]*background:\s*#191919/s)
