+++ packages/public-mascot/package.json
@@ -1,0 +1,13 @@
+{
+  "name": "@biu/public-mascot",
+  "version": "0.1.0",
+  "private": true,
+  "type": "module",
+  "description": "Mascot 共享 UI（public：不是 Cordis 插件；壳与 chat 共用）",
+  "exports": {
+    ".": "./src/web/index.ts"
+  },
+  "peerDependencies": {
+    "react": "^19.1.1"
+  }
+}
