+++ packages/public-mascot/src/web/NOTICE.md
