+++ packages/public-mascot/src/web/brand-mascot.tsx
