+++ packages/public-mascot/src/web/dance-stage.test.ts
