+++ packages/public-mascot/src/web/dance-stage.tsx
