+++ packages/public-mascot/src/web/grok-bot-avatar.tsx
