+++ packages/public-mascot/src/web/grok-bot-loader.ts
