+++ packages/public-mascot/src/web/grok-bot-types.ts
