+++ packages/public-mascot/src/web/mascot-dance.test.ts
