+++ packages/public-mascot/src/web/mascot-dance.ts
