+++ packages/public-mascot/src/web/session-mascot.test.ts
