+++ packages/public-mascot/src/web/sidebar-mascot.tsx
