+++ packages/public-mascot/src/web/static-mascot-mark.tsx
