+++ packages/public-ui/src/bool-box.tsx
@@ -1,0 +1,15 @@
+import type { ReactNode } from 'react'
+
+export function BoolBox({
+  on,
+  locked,
+  children,
+}: {
+  on: boolean
+  locked?: boolean
+  children?: ReactNode
+}) {
+  return (
+    <span className={`fsdb-boolbox${on ? ' is-on' : ''}${locked ? ' is-locked' : ''}`}>{children}</span>
+  )
+}
