+++ packages/web-app-shell/src/web/brand-corner.test.ts
@@ -22,7 +22,7 @@ test('brand mascot lives at the corner; sidebar head is title plus collapse', ()
   assert.match(chat, /openOverlayComposer/)
   assert.match(chat, /sessionView.load\(item.id/)
   assert.doesNotMatch(chat, /SidebarBrandMascot/)
-  const mascot = readFileSync(resolve(import.meta.dirname, '../../../web-mascot/src/web/brand-mascot.tsx'), 'utf8')
+  const mascot = readFileSync(resolve(import.meta.dirname, '../../../public-mascot/src/web/brand-mascot.tsx'), 'utf8')
   assert.match(mascot, /brand-agent-menu/)
   assert.match(mascot, /data-dock-tip/)
   assert.match(mascot, /BrandAgentMenu/)
