+++ packages/web-app-shell/src/web/session-inspector.tsx
@@ -25,7 +25,7 @@ import type { SlotsService } from '@biu/web-slots'
 import { inspectorTabFromEvent, requestInspectorAction } from './chat-overlay.ts'
 import { getInspectorCaption, getInspectorCaptionVersion, subscribeInspectorCaptions } from './inspector-captions.ts'
 import { inspectorPanelMatches, inspectorViewProps, nextRepeatableTabId, resolveInspectorTab, slotTabId } from './inspector-panels.ts'
-import { SidebarMascot, resolveSessionMascot } from '@biu/web-mascot'
+import { SidebarMascot, resolveSessionMascot } from '@biu/public-mascot'
 
 function captionTableIcon(icon?: string) {
   const name = (icon ?? '').trim().toLowerCase()
